Moving BaseWindow to it's own file for clarity's sake
